+++ src/cluster/tests.rs
@@ -66,7 +66,7 @@ fn single_node_cluster() {
     let new_state = ConsensusData {
         height: 100,
         round: 1,
-        step: SignedMsgType::Proposal as u8,
+        step: SignedMsgType::Proposal,
     };
 
     let result = cluster.replicate_state(&new_state);
@@ -122,7 +122,7 @@ fn three_node_cluster_basic() {
     let new_state = ConsensusData {
         height: 200,
         round: 2,
-        step: SignedMsgType::Prevote as u8,
+        step: SignedMsgType::Prevote,
     };
 
     let result = leader.replicate_state(&new_state);
